Move remote storage dependencies into new module
