+++ gnd/src/main/java/com/google/android/gnd/GndApplicationModule.java
@@ -63,44 +63,4 @@ abstract class GndApplicationModule {
         .getApplicationContext()
         .getSharedPreferences(Config.SHARED_PREFS_NAME, Config.SHARED_PREFS_MODE);
   }
-
-  @Provides
-  static FirebaseFirestoreSettings firebaseFirestoreSettings() {
-    return new FirebaseFirestoreSettings.Builder()
-        .setPersistenceEnabled(Config.FIRESTORE_PERSISTENCE_ENABLED)
-        .build();
-  }
-
-  @Provides
-  @Singleton
-  static FirebaseFirestore firebaseFirestore(FirebaseFirestoreSettings settings) {
-    FirebaseFirestore firestore = FirebaseFirestore.getInstance();
-    firestore.setFirestoreSettings(settings);
-    FirebaseFirestore.setLoggingEnabled(Config.FIRESTORE_LOGGING_ENABLED);
-    return firestore;
-  }
-
-  /** Returns a reference to the default Storage bucket. */
-  @Provides
-  @Singleton
-  static StorageReference firebaseStorageReference() {
-    return FirebaseStorage.getInstance().getReference();
-  }
-
-  /** Provides the Firestore implementation of remote data store. */
-  @Binds
-  @Singleton
-  abstract RemoteDataStore remoteDataStore(FirestoreDataStore ds);
-
-  /** Provides the Firestore implementation of offline unique id generation. */
-  @Binds
-  @Singleton
-  abstract OfflineUuidGenerator offlineUuidGenerator(FirestoreUuidGenerator uuidGenerator);
-
-  /** Provides the Firestore implementation of remote storage manager. */
-  @Binds
-  @Singleton
-  abstract RemoteStorageManager remoteStorageManager(FirestoreStorageManager fsm);
-
-
 }
